Chapter 1: Case Study and Chapter End Questions
The following work is completed as part of my personal advancement for the IST 687 - Introduction to Data Science course, taught by Christopher Dunham, in pursuit of my Master of Science in Applied Human-Centered AI at Syracuse University. The questions completed are outlined at the end of Chapter 1 in “Data Science for Business with R” by Jeffrey S. Saltz and Jeffrey M. Stanton.